More renames and comments
