+++ packages/client/src/__tests__/batching.test.ts
@@ -1,7 +1,7 @@
 import { blog } from '../fixtures/blog'
 import { getDMMF } from '../generation/getDMMF'
 import { DMMFClass, makeDocument } from '../runtime'
-import { MergedExtensionsList } from '../runtime/core/extensions/ExtensionsList'
+import { MergedExtensionsList } from '../runtime/core/extensions/MergedExtensionsList'
 import { RequestHandler } from '../runtime/RequestHandler'
 
 describe('batching', () => {
