+++ packages/client/src/__tests__/or.test.ts
@@ -3,7 +3,7 @@ import stripAnsi from 'strip-ansi'
 import { enums } from '../fixtures/enums'
 import { getDMMF } from '../generation/getDMMF'
 import { DMMFClass, makeDocument } from '../runtime'
-import { MergedExtensionsList } from '../runtime/core/extensions/ExtensionsList'
+import { MergedExtensionsList } from '../runtime/core/extensions/MergedExtensionsList'
 
 describe('at least one validation', () => {
   let dmmf
