+++ packages/client/src/__tests__/query/decimal.test.ts
@@ -2,7 +2,7 @@ import { Decimal } from 'decimal.js'
 
 import { getDMMF } from '../../generation/getDMMF'
 import { DMMFClass, makeDocument } from '../../runtime'
-import { MergedExtensionsList } from '../../runtime/core/extensions/ExtensionsList'
+import { MergedExtensionsList } from '../../runtime/core/extensions/MergedExtensionsList'
 
 const datamodel = /* Prisma */ `
     datasource my_db {
