+++ packages/client/src/__tests__/undefined-vs-null.test.ts
@@ -2,7 +2,7 @@ import stripAnsi from 'strip-ansi'
 
 import { blog } from '../fixtures/blog'
 import { getDMMF } from '../generation/getDMMF'
-import { MergedExtensionsList } from '../runtime/core/extensions/ExtensionsList'
+import { MergedExtensionsList } from '../runtime/core/extensions/MergedExtensionsList'
 import { DMMFHelper } from '../runtime/dmmf'
 import { makeDocument } from '../runtime/query'
